fix(recipes): address query review edge cases
Restore scalar activityIdIn mappings, preserve unfiltered query semantics, and track parenthesized and field assignments across scopes. Update regression expectations and coverage for count result types.

+++ code-conversion/recipes/src/test/java/io/camunda/migration/code/recipes/delegate/JavaDelegateWithClientQueryRecipesTest.java
@@ -102,7 +102,7 @@ public class JavaDelegateWithClientQueryRecipesTest implements RewriteTest {
                             .send()
                             .join()
                             .page()
-                            .totalItems() % 2 == 0;
+                            .totalItems().intValue() % 2 == 0;
 
                     long activeCount = camundaClient
                             .newProcessInstanceSearchRequest()
